+++ srcs/backend/prisma/schema.prisma
@@ -1,59 +1,87 @@
 generator client {
-  provider = "prisma-client-js"
+    provider = "prisma-client-js"
 }
 
 datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
+    provider = "postgresql"
+    url      = env("DATABASE_URL")
 }
 
 model User {
-  id             String    @id @default(cuid())
-  authid         String    @unique
-  createdAt      DateTime  @default(now())
-  updatedAt      DateTime  @updatedAt
-  profile        Profile?
-  auth           Auth      @relation(fields: [authid], references: [id])
-  blockedProfile Profile[] @relation("blockedProfile")
+    userID              String      @id @default(cuid())
+    profile             Profile?
+    auth                Auth?
+    createdAt           DateTime    @default(now())
+    updatedAt           DateTime    @updatedAt
 }
 
 model Auth {
-  id           String @id @default(cuid())
-  accessToken  String @unique @default("")
-  refreshToken String @unique @default("")
-  bearerToken  String @unique @default("")
-  user         User?
-}
-
-model Stats {
-  id        String   @id @default(cuid())
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-  profile   Profile?
-}
-
-model Ranking {
-  id      String   @id @default(cuid())
-  win     Int      @default(0)
-  lose    Int      @default(0)
-  streak  Int      @default(0)
-  date    DateTime @default(now())
-  profile Profile?
+    authID              String      @id @default(cuid())
+    user                User        @relation(fields: [authID], references: [userID])
+    accessToken         String      @unique @default("")
+    refreshToken        String      @unique @default("")
+    bearerToken         String      @unique @default("")
 }
 
 model Profile {
-  id          String   @id @default(cuid())
-  username    String   @unique
-  displayName String
-  status      String   @default("ACTIVE")
-  image       String   @default("https://www.gravatar.com/avatar/00000000000000000000000000000000?d=mp&f=y")
-  userId      String   @unique
-  statsId     String   @unique
-  rankingId   String   @unique
-  createdAt   DateTime @default(now())
-  updatedAt   DateTime @updatedAt
-  user        User     @relation(fields: [userId], references: [id])
-  stats       Stats    @relation(fields: [statsId], references: [id])
-  ranking     Ranking  @relation(fields: [rankingId], references: [id])
-  blockedBy   User[]   @relation("blockedProfile")
+    profileID           String      @id @default(cuid())
+    user                User        @relation(fields: [profileID], references: [userID])
+    login42             String
+    username            String      @unique
+    online              Boolean     @default(true)
+    coalition           String
+    imagePath           String
+    authentificator     Boolean
+    jwtToken            Json
+    rightGame           Game[]      @relation(name: "rightPlayer")
+    leftGame            Game[]      @relation(name: "leftPlayer")
+    friendReqSent       Friend[]    @relation(name: "sent")
+    friendReqRece       Friend[]    @relation(name: "received")
+    blocked             Block[]     @relation(name: "blocker")
+    blockedBy           Block[]     @relation(name: "blocked")
+    channelOwned        Channel[]   @relation(name: "owner")
+    channels            Channel[]
+    createdAt           DateTime    @default(now())
+    updatedAt           DateTime    @updatedAt
+}
+
+model Game{
+    gameID              String      @id @default(cuid())
+    leftPlayer          Profile     @relation(name: "leftPlayer", fields: [leftPlayerID], references: [profileID])
+    leftPlayerID        String
+    leftPlayerScore     Int         @default(0)
+    rightPlayer         Profile     @relation(name: "rightPlayer", fields: [rightPlayerID], references: [profileID])
+    rightPlayerID       String
+    rightPlayerScore    Int         @default(0)
+    active              Boolean     @default(true)
+    winner              String      @default("Undefined")
+}
+
+model Friend{
+    friendID            String      @id @default(cuid())
+    senderProfile       Profile     @relation(name: "sent", fields: [senderID], references: [profileID])
+    senderID            String
+    receiverProfile     Profile     @relation(name: "received" ,fields: [receiverID], references: [profileID])
+    receiverID          String
+    status              String      @default("Pending")
+}
+
+model Block{
+    blockID             String      @id @default(cuid())
+    blockerProfile      Profile     @relation(name: "blocker", fields: [blockerID], references: [profileID])
+    blockerID           String
+    blockedProfile      Profile     @relation(name: "blocked", fields: [blockedID], references: [profileID])
+    blockedID           String
+    reason              String      @default("None")
+}
+
+model Channel{
+    channelID           String      @id @default(cuid())
+    Name                String      @unique @default("Name not initialised")
+    private             Boolean     @default(false)
+    password            String
+    owner               Profile     @relation(name: "owner", fields: [ownerID], references: [profileID])
+    ownerID             String
+    users               Profile[]
+    admin               Boolean     @default(false)
 }
