+++ srcs/backend/src/app.controller.spec.ts
@@ -3,20 +3,20 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 
 describe('AppController', () => {
-  let appController: AppController;
+//   let appController: AppController;
 
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
-      controllers: [AppController],
-      providers: [AppService],
-    }).compile();
+//   beforeEach(async () => {
+//     const app: TestingModule = await Test.createTestingModule({
+//       controllers: [AppController],
+//       providers: [AppService],
+//     }).compile();
 
-    appController = app.get<AppController>(AppController);
-  });
+//     appController = app.get<AppController>(AppController);
+//   });
 
-  describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(appController.getHello()).toBe('Hello World!');
-    });
-  });
+//   describe('root', () => {
+//     it('should return "Hello World!"', () => {
+//       expect(appController.getHello()).toBe('Hello World!');
+//     });
+//   });
 });
