+++ srcs/backend/src/app.controller.ts
@@ -1,12 +1,12 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
+// import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
-  constructor(private readonly appService: AppService) {}
+//   constructor(private readonly appService: AppService) {}
 
-  @Get()
-  getHello(): object {
-    return this.appService.getHello();
-  }
+//   @Get()
+//   getHello(): object {
+//     return this.appService.getHello();
+//   }
 }
