+++ srcs/backend/src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AuthModule } from './auth/auth.module';
-import { UsersModule } from './users/users.module';
+// import { AuthModule } from './auth/auth.module';
+// import { UsersModule } from './users/users.module';
 
 @Module({
-  imports: [AuthModule, UsersModule],
+//   imports: [AuthModule, UsersModule],
   controllers: [AppController],
   providers: [AppService],
 })
