+++ srcs/backend/src/app.service.ts
@@ -7,11 +7,11 @@ export class AppService {
   // retun a json object
   // get  all users from the database
 
-  getHello(): object {
+//   getHello(): object {
     //prisma create a user
-    const data = {
-      username: 'test',
-    };
-    return { message: 'Hello World!' };
-  }
+    // const data = {
+    //   username: 'test',
+    // };
+    // return { message: 'Hello World!' };
+//   }
 }
