+++ srcs/backend/src/auth/auth.controller.spec.ts
@@ -1,20 +1,20 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+// import { Test, TestingModule } from '@nestjs/testing';
+// import { AuthController } from './auth.controller';
+// import { AuthService } from './auth.service';
 
-describe('AuthController', () => {
-  let controller: AuthController;
+// describe('AuthController', () => {
+//   let controller: AuthController;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [AuthController],
-      providers: [AuthService],
-    }).compile();
+//   beforeEach(async () => {
+//     const module: TestingModule = await Test.createTestingModule({
+//       controllers: [AuthController],
+//       providers: [AuthService],
+//     }).compile();
 
-    controller = module.get<AuthController>(AuthController);
-  });
+//     controller = module.get<AuthController>(AuthController);
+//   });
 
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+//   it('should be defined', () => {
+//     expect(controller).toBeDefined();
+//   });
+// });
