+++ srcs/backend/src/auth/dto/create-auth.dto.ts
@@ -1,7 +1,7 @@
-export class CreateAuthDto {
-  username: string;
-  displayName: string;
-  constructor(partial: Partial<CreateAuthDto>) {
-    Object.assign(this, partial);
-  }
-}
+// export class CreateAuthDto {
+//   username: string;
+//   displayName: string;
+//   constructor(partial: Partial<CreateAuthDto>) {
+//     Object.assign(this, partial);
+//   }
+// }
