+++ srcs/backend/src/auth/entities/auth.entity.ts
@@ -1,1 +1,1 @@
-export class Auth {}
+// export class Auth {}
