+++ srcs/backend/src/users/dto/update-user.dto.ts
@@ -1,4 +1,4 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateUserDto } from './create-user.dto';
+// import { PartialType } from '@nestjs/mapped-types';
+// import { CreateUserDto } from './create-user.dto';
 
-export class UpdateUserDto extends PartialType(CreateUserDto) {}
+// export class UpdateUserDto extends PartialType(CreateUserDto) {}
