+++ srcs/backend/src/users/entities/user.entity.ts
@@ -1,1 +1,1 @@
-export class User {}
+// export class User {}
