+++ srcs/backend/src/users/users.controller.ts
@@ -1,42 +1,42 @@
-import {
-  Controller,
-  Get,
-  Post,
-  Body,
-  Patch,
-  Param,
-  Delete,
-} from '@nestjs/common';
-import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+// import {
+//   Controller,
+//   Get,
+//   Post,
+//   Body,
+//   Patch,
+//   Param,
+//   Delete,
+// } from '@nestjs/common';
+// import { UsersService } from './users.service';
+// import { CreateUserDto } from './dto/create-user.dto';
+// import { UpdateUserDto } from './dto/update-user.dto';
 
-@Controller('users')
-export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+// @Controller('users')
+// export class UsersController {
+//   constructor(private readonly usersService: UsersService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
-  }
+//   @Post()
+//   create(@Body() createUserDto: CreateUserDto) {
+//     return this.usersService.create(createUserDto);
+//   }
 
-  @Get()
-  findAll() {
-    return this.usersService.findAll();
-  }
-  // put this behind a guard so only logged in users can access it
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
-  }
+//   @Get()
+//   findAll() {
+//     return this.usersService.findAll();
+//   }
+//   // put this behind a guard so only logged in users can access it
+//   @Get(':id')
+//   findOne(@Param('id') id: string) {
+//     return this.usersService.findOne(id);
+//   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
-  }
+//   @Patch(':id')
+//   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+//     return this.usersService.update(+id, updateUserDto);
+//   }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
-  }
-}
+//   @Delete(':id')
+//   remove(@Param('id') id: string) {
+//     return this.usersService.remove(+id);
+//   }
+// }
