+++ srcs/backend/src/users/users.service.ts
@@ -1,28 +1,28 @@
-import { Injectable } from '@nestjs/common';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { prisma } from '../main';
-@Injectable()
-export class UsersService {
-  create(createUserDto: CreateUserDto) {
-    return 'This action adds a new user';
-  }
+// import { Injectable } from '@nestjs/common';
+// import { CreateUserDto } from './dto/create-user.dto';
+// import { UpdateUserDto } from './dto/update-user.dto';
+// import { prisma } from '../main';
+// @Injectable()
+// export class UsersService {
+//   create(createUserDto: CreateUserDto) {
+//     return 'This action adds a new user';
+//   }
 
-  findAll() {
-    return prisma.profile.findMany();
-  }
-  async findOne(username: string) {
-    //return user profile
-    const data = await prisma.profile.findUnique({
-      where: { username: username },
-    });
-  }
+//   findAll() {
+//     return prisma.profile.findMany();
+//   }
+//   async findOne(username: string) {
+//     //return user profile
+//     const data = await prisma.profile.findUnique({
+//       where: { username: username },
+//     });
+//   }
 
-  update(id: number, updateUserDto: UpdateUserDto) {
-    return `This action updates a #${id} user`;
-  }
+//   update(id: number, updateUserDto: UpdateUserDto) {
+//     return `This action updates a #${id} user`;
+//   }
 
-  remove(id: number) {
-    return `This action removes a #${id} user`;
-  }
-}
+//   remove(id: number) {
+//     return `This action removes a #${id} user`;
+//   }
+// }
