+++ srcs/frontend/src/views/Watch/Watch.tsx
@@ -1,4 +1,5 @@
 import Chat from './Chat/Chat'
+
 export default function Watch() {
   // canvas should be  replace by the actual game and the chat should be replaced by the actual chat
   return (
